refactor(cli): isolate allocation audit telemetry
