+++ crates/bitnet-cli/src/main.rs
@@ -18,66 +18,14 @@ use candle_core::{DType, IndexOp};
 use clap::{CommandFactory, Parser, Subcommand};
 use clap_complete::{Shell, generate};
 use console::style;
-use std::alloc::{GlobalAlloc, Layout, System};
 use std::io;
-use std::sync::atomic::{AtomicBool, AtomicU64, Ordering};
 use tracing::{debug, error, info, warn};
 
 #[global_allocator]
-static ALLOCATION_AUDIT_ALLOCATOR: AllocationAuditAllocator = AllocationAuditAllocator;
+static ALLOCATION_AUDIT_ALLOCATOR: allocation_audit::AllocationAuditAllocator =
+    allocation_audit::AllocationAuditAllocator;
 
-static ALLOCATION_AUDIT_ENABLED: AtomicBool = AtomicBool::new(false);
-static ALLOCATION_AUDIT_ALLOC_COUNT: AtomicU64 = AtomicU64::new(0);
-static ALLOCATION_AUDIT_ALLOC_BYTES: AtomicU64 = AtomicU64::new(0);
-static ALLOCATION_AUDIT_DEALLOC_COUNT: AtomicU64 = AtomicU64::new(0);
-static ALLOCATION_AUDIT_DEALLOC_BYTES: AtomicU64 = AtomicU64::new(0);
-
-struct AllocationAuditAllocator;
-
-unsafe impl GlobalAlloc for AllocationAuditAllocator {
-    unsafe fn alloc(&self, layout: Layout) -> *mut u8 {
-        let ptr = unsafe { System.alloc(layout) };
-        if !ptr.is_null() && ALLOCATION_AUDIT_ENABLED.load(Ordering::Relaxed) {
-            record_allocation_audit_alloc(layout.size());
-        }
-        ptr
-    }
-
-    unsafe fn alloc_zeroed(&self, layout: Layout) -> *mut u8 {
-        let ptr = unsafe { System.alloc_zeroed(layout) };
-        if !ptr.is_null() && ALLOCATION_AUDIT_ENABLED.load(Ordering::Relaxed) {
-            record_allocation_audit_alloc(layout.size());
-        }
-        ptr
-    }
-
-    unsafe fn dealloc(&self, ptr: *mut u8, layout: Layout) {
-        if ALLOCATION_AUDIT_ENABLED.load(Ordering::Relaxed) {
-            record_allocation_audit_dealloc(layout.size());
-        }
-        unsafe { System.dealloc(ptr, layout) };
-    }
-
-    unsafe fn realloc(&self, ptr: *mut u8, layout: Layout, new_size: usize) -> *mut u8 {
-        let new_ptr = unsafe { System.realloc(ptr, layout, new_size) };
-        if !new_ptr.is_null() && ALLOCATION_AUDIT_ENABLED.load(Ordering::Relaxed) {
-            record_allocation_audit_dealloc(layout.size());
-            record_allocation_audit_alloc(new_size);
-        }
-        new_ptr
-    }
-}
-
-fn record_allocation_audit_alloc(size: usize) {
-    ALLOCATION_AUDIT_ALLOC_COUNT.fetch_add(1, Ordering::Relaxed);
-    ALLOCATION_AUDIT_ALLOC_BYTES.fetch_add(size as u64, Ordering::Relaxed);
-}
-
-fn record_allocation_audit_dealloc(size: usize) {
-    ALLOCATION_AUDIT_DEALLOC_COUNT.fetch_add(1, Ordering::Relaxed);
-    ALLOCATION_AUDIT_DEALLOC_BYTES.fetch_add(size as u64, Ordering::Relaxed);
-}
-
+mod allocation_audit;
 #[cfg(feature = "full-cli")]
 mod commands;
 mod config;
@@ -94,6 +42,15 @@ mod score;
 mod simple_generation;
 pub mod tokenizer_discovery;
 
+use allocation_audit::{
+    AllocationAuditGuard, AllocationAuditSnapshot, allocation_bytes_delta_json,
+    allocation_count_delta_json, allocation_samples_json, mean_alloc_bytes, mean_alloc_count,
+};
+#[cfg(feature = "full-cli")]
+use allocation_audit::{
+    WarmSessionPromptAllocationAudit, warm_session_aggregate_allocation_audit_json,
+    warm_session_prompt_allocation_audit_json,
+};
 use exit::*;
 
 /// Build the CLI command for external use (e.g., in tests)
@@ -11384,52 +11341,6 @@ fn rounded_ms(ms: f64) -> f64 {
     (ms * 1000.0).round() / 1000.0
 }
 
-#[derive(Clone, Copy, Debug, Default, PartialEq, Eq)]
-struct AllocationAuditSnapshot {
-    alloc_count: u64,
-    alloc_bytes: u64,
-    dealloc_count: u64,
-    dealloc_bytes: u64,
-}
-
-impl AllocationAuditSnapshot {
-    fn current() -> Self {
-        Self {
-            alloc_count: ALLOCATION_AUDIT_ALLOC_COUNT.load(Ordering::Relaxed),
-            alloc_bytes: ALLOCATION_AUDIT_ALLOC_BYTES.load(Ordering::Relaxed),
-            dealloc_count: ALLOCATION_AUDIT_DEALLOC_COUNT.load(Ordering::Relaxed),
-            dealloc_bytes: ALLOCATION_AUDIT_DEALLOC_BYTES.load(Ordering::Relaxed),
-        }
-    }
-
-    fn delta_since(start: Self) -> Self {
-        let current = Self::current();
-        Self {
-            alloc_count: current.alloc_count.saturating_sub(start.alloc_count),
-            alloc_bytes: current.alloc_bytes.saturating_sub(start.alloc_bytes),
-            dealloc_count: current.dealloc_count.saturating_sub(start.dealloc_count),
-            dealloc_bytes: current.dealloc_bytes.saturating_sub(start.dealloc_bytes),
-        }
-    }
-}
-
-struct AllocationAuditGuard {
-    previous: bool,
-}
-
-impl AllocationAuditGuard {
-    fn enable(enabled: bool) -> Self {
-        let previous = ALLOCATION_AUDIT_ENABLED.swap(enabled, Ordering::Relaxed);
-        Self { previous }
-    }
-}
-
-impl Drop for AllocationAuditGuard {
-    fn drop(&mut self) {
-        ALLOCATION_AUDIT_ENABLED.store(self.previous, Ordering::Relaxed);
-    }
-}
-
 fn timing_samples_json(samples: &[f64]) -> serde_json::Value {
     if samples.is_empty() {
         return serde_json::json!({
@@ -11493,319 +11404,6 @@ fn tokens_per_second_json(tokens: usize, elapsed_ms: f64) -> serde_json::Value {
     } else {
         serde_json::json!(rounded_ms(tokens as f64 / (elapsed_ms / 1000.0)))
     }
-}
-
-fn allocation_samples_json(samples: &[AllocationAuditSnapshot]) -> serde_json::Value {
-    if samples.is_empty() {
-        return serde_json::json!({
-            "count": 0,
-            "alloc_count_total": 0,
-            "alloc_bytes_total": 0,
-            "dealloc_count_total": 0,
-            "dealloc_bytes_total": 0,
-            "net_bytes_total": 0,
-            "mean_alloc_count_per_token": serde_json::Value::Null,
-            "mean_alloc_bytes_per_token": serde_json::Value::Null,
-            "max_alloc_count_per_token": serde_json::Value::Null,
-            "max_alloc_bytes_per_token": serde_json::Value::Null,
-        });
-    }
-
-    let total_alloc_count = samples.iter().map(|sample| sample.alloc_count).sum::<u64>();
-    let total_alloc_bytes = samples.iter().map(|sample| sample.alloc_bytes).sum::<u64>();
-    let total_dealloc_count = samples.iter().map(|sample| sample.dealloc_count).sum::<u64>();
-    let total_dealloc_bytes = samples.iter().map(|sample| sample.dealloc_bytes).sum::<u64>();
-    let max_alloc_count = samples.iter().map(|sample| sample.alloc_count).max().unwrap_or(0);
-    let max_alloc_bytes = samples.iter().map(|sample| sample.alloc_bytes).max().unwrap_or(0);
-    let count = samples.len() as f64;
-
-    let net_bytes_total = total_alloc_bytes as i64 - total_dealloc_bytes as i64;
-
-    serde_json::json!({
-        "count": samples.len(),
-        "alloc_count_total": total_alloc_count,
-        "alloc_bytes_total": total_alloc_bytes,
-        "dealloc_count_total": total_dealloc_count,
-        "dealloc_bytes_total": total_dealloc_bytes,
-        "net_bytes_total": net_bytes_total,
-        "mean_alloc_count_per_token": rounded_ms(total_alloc_count as f64 / count),
-        "mean_alloc_bytes_per_token": rounded_ms(total_alloc_bytes as f64 / count),
-        "max_alloc_count_per_token": max_alloc_count,
-        "max_alloc_bytes_per_token": max_alloc_bytes,
-    })
-}
-
-#[cfg(feature = "full-cli")]
-struct WarmSessionPromptAllocationAudit<'a> {
-    enabled: bool,
-    requested_backend: &'a str,
-    prompt_tokenize: AllocationAuditSnapshot,
-    prompt_setup: AllocationAuditSnapshot,
-    prompt_prefill: &'a [AllocationAuditSnapshot],
-    decode_total: &'a [AllocationAuditSnapshot],
-    embed: &'a [AllocationAuditSnapshot],
-    forward: &'a [AllocationAuditSnapshot],
-    logits: &'a [AllocationAuditSnapshot],
-    sample: &'a [AllocationAuditSnapshot],
-    token_vector_update: &'a [AllocationAuditSnapshot],
-    token_decode: &'a [AllocationAuditSnapshot],
-    stop_tail_update: &'a [AllocationAuditSnapshot],
-    receipt_construction: AllocationAuditSnapshot,
-}
-
-#[cfg(feature = "full-cli")]
-fn warm_session_prompt_allocation_audit_json(
-    audit: WarmSessionPromptAllocationAudit<'_>,
-) -> serde_json::Value {
-    if !audit.enabled {
-        return serde_json::json!({
-            "enabled": false,
-            "method": "not_requested",
-            "scope": "not_requested",
-        });
-    }
-
-    let prompt_tokenize = std::slice::from_ref(&audit.prompt_tokenize);
-    let prompt_setup = std::slice::from_ref(&audit.prompt_setup);
-    let receipt_construction = std::slice::from_ref(&audit.receipt_construction);
-    let mut hotspots = vec![
-        allocation_hotspot("prompt_tokenize", prompt_tokenize),
-        allocation_hotspot("prompt_setup", prompt_setup),
-        allocation_hotspot("prompt_prefill", audit.prompt_prefill),
-        allocation_hotspot("decode_total", audit.decode_total),
-        allocation_hotspot("model.embed", audit.embed),
-        allocation_hotspot("model.forward", audit.forward),
-        allocation_hotspot("model.logits_and_extract", audit.logits),
-        allocation_hotspot("sampler.sample", audit.sample),
-        allocation_hotspot("token_vector_updates", audit.token_vector_update),
-        allocation_hotspot("tokenizer.decode", audit.token_decode),
-        allocation_hotspot("stop_tail_updates", audit.stop_tail_update),
-        allocation_hotspot("receipt_construction", receipt_construction),
-    ];
-    hotspots.retain(|hotspot| hotspot.alloc_count > 0 || hotspot.alloc_bytes > 0);
-    hotspots.sort_by(|left, right| {
-        right
-            .alloc_bytes
-            .cmp(&left.alloc_bytes)
-            .then_with(|| right.alloc_count.cmp(&left.alloc_count))
-            .then_with(|| left.component.cmp(right.component))
-    });
-
-    serde_json::json!({
-        "enabled": true,
-        "method": "process_global_allocator_counter_delta",
-        "scope": warm_session_allocation_scope(audit.requested_backend),
-        "claim_scope": "allocation counter deltas for this prompt/profile only; sampling scratch cleanup is scoped and no broad performance improvement is claimed",
-        "optimization_deferred": false,
-        "unavoidable_candidates_named_before_optimization": true,
-        "ranked_hotspots": hotspots.iter().map(AllocationHotspot::to_json).collect::<Vec<_>>(),
-        "unavoidable_candidates": [
-            "model.embed/model.forward/model.logits tensor outputs from the current dense Qwen CPU execution path",
-            "tokenizer.decode allocation for per-token text and stop-tail checks",
-            "receipt construction outside the decode hot loop",
-            "prompt token vector growth is controlled by reusable session buffers; model tensor outputs remain the dominant allocation source"
-        ],
-        "instrumentation_included": [
-            "prompt_tokenize",
-            "prompt_setup",
-            "prompt_prefill_step",
-            "decode_step_total",
-            "model.embed",
-            "model.forward",
-            "model.logits_and_extract",
-            "sampler.sample",
-            "token_vector_updates",
-            "tokenizer.decode",
-            "stop_tail_updates",
-            "receipt_construction"
-        ],
-        "instrumentation_excluded": [
-            "model_load",
-            "tokenizer_load",
-            "aggregate_receipt_serialization",
-            "OS allocator reuse and fragmentation outside counter deltas"
-        ],
-        "prompt_tokenize": allocation_samples_json(prompt_tokenize),
-        "prompt_setup": allocation_samples_json(prompt_setup),
-        "prompt_prefill": allocation_samples_json(audit.prompt_prefill),
-        "decode": {
-            "total": allocation_samples_json(audit.decode_total),
-            "steady_state": allocation_samples_json(audit.decode_total.get(1..).unwrap_or(&[])),
-            "embed": allocation_samples_json(audit.embed),
-            "forward": allocation_samples_json(audit.forward),
-            "logits": allocation_samples_json(audit.logits),
-            "sample": allocation_samples_json(audit.sample),
-            "token_vector_update": allocation_samples_json(audit.token_vector_update),
-            "token_decode": allocation_samples_json(audit.token_decode),
-            "stop_tail_update": allocation_samples_json(audit.stop_tail_update),
-        },
-        "receipt_construction": allocation_samples_json(receipt_construction),
-    })
-}
-
-#[cfg(feature = "full-cli")]
-fn warm_session_aggregate_allocation_audit_json(
-    enabled: bool,
-    requested_backend: &str,
-    prompt_summaries: &[serde_json::Value],
-) -> serde_json::Value {
-    if !enabled {
-        return serde_json::json!({
-            "enabled": false,
-            "method": "not_requested",
-            "scope": "not_requested",
-        });
-    }
-
-    let mut totals = std::collections::BTreeMap::<String, (u64, u64)>::new();
-    for prompt in prompt_summaries {
-        let Some(hotspots) = prompt["allocation_audit"]["ranked_hotspots"].as_array() else {
-            continue;
-        };
-        for hotspot in hotspots {
-            let Some(component) = hotspot["component"].as_str() else {
-                continue;
-            };
-            let entry = totals.entry(component.to_string()).or_default();
-            entry.0 += hotspot["alloc_count"].as_u64().unwrap_or_default();
-            entry.1 += hotspot["alloc_bytes"].as_u64().unwrap_or_default();
-        }
-    }
-    let mut ranked = totals
-        .into_iter()
-        .map(|(component, (alloc_count, alloc_bytes))| {
-            serde_json::json!({
-                "component": component,
-                "alloc_count": alloc_count,
-                "alloc_bytes": alloc_bytes,
-            })
-        })
-        .collect::<Vec<_>>();
-    ranked.sort_by(|left, right| {
-        right["alloc_bytes"]
-            .as_u64()
-            .unwrap_or_default()
-            .cmp(&left["alloc_bytes"].as_u64().unwrap_or_default())
-            .then_with(|| {
-                right["alloc_count"]
-                    .as_u64()
-                    .unwrap_or_default()
-                    .cmp(&left["alloc_count"].as_u64().unwrap_or_default())
-            })
-            .then_with(|| {
-                left["component"]
-                    .as_str()
-                    .unwrap_or_default()
-                    .cmp(right["component"].as_str().unwrap_or_default())
-            })
-    });
-
-    serde_json::json!({
-        "enabled": true,
-        "method": "process_global_allocator_counter_delta",
-        "scope": warm_session_allocation_scope(requested_backend),
-        "claim_scope": "aggregate of prompt-level allocation counter deltas; sampling scratch cleanup is scoped and no broad performance improvement is claimed",
-        "prompt_count": prompt_summaries.len(),
-        "ranked_hotspots": ranked,
-        "optimization_deferred": false,
-    })
-}
-
-#[cfg(feature = "full-cli")]
-fn warm_session_allocation_scope(requested_backend: &str) -> &'static str {
-    match requested_backend.trim().to_ascii_lowercase().as_str() {
-        "cpu" => "selected generic CPU SLM warm-session prompt hot path",
-        "apple-m3-air-cpu-neon" => {
-            "selected Apple M3 Air CPU/NEON SLM warm-session prompt hot path"
-        }
-        _ => "selected Apple M4 CPU/NEON SLM warm-session prompt hot path",
-    }
-}
-
-#[cfg(feature = "full-cli")]
-struct AllocationHotspot {
-    component: &'static str,
-    alloc_count: u64,
-    alloc_bytes: u64,
-}
-
-#[cfg(feature = "full-cli")]
-impl AllocationHotspot {
-    fn to_json(&self) -> serde_json::Value {
-        serde_json::json!({
-            "component": self.component,
-            "alloc_count": self.alloc_count,
-            "alloc_bytes": self.alloc_bytes,
-        })
-    }
-}
-
-#[cfg(feature = "full-cli")]
-fn allocation_hotspot(
-    component: &'static str,
-    samples: &[AllocationAuditSnapshot],
-) -> AllocationHotspot {
-    AllocationHotspot {
-        component,
-        alloc_count: samples.iter().map(|sample| sample.alloc_count).sum(),
-        alloc_bytes: samples.iter().map(|sample| sample.alloc_bytes).sum(),
-    }
-}
-
-fn allocation_count_delta_json(samples: &[AllocationAuditSnapshot]) -> serde_json::Value {
-    if samples.is_empty() {
-        return serde_json::json!({
-            "count": 0,
-            "total": 0,
-            "mean_per_token": serde_json::Value::Null,
-            "max_per_token": serde_json::Value::Null,
-        });
-    }
-
-    let total = samples.iter().map(|sample| sample.alloc_count).sum::<u64>();
-    let max = samples.iter().map(|sample| sample.alloc_count).max().unwrap_or(0);
-
-    serde_json::json!({
-        "count": samples.len(),
-        "total": total,
-        "mean_per_token": rounded_ms(total as f64 / samples.len() as f64),
-        "max_per_token": max,
-    })
-}
-
-fn allocation_bytes_delta_json(samples: &[AllocationAuditSnapshot]) -> serde_json::Value {
-    if samples.is_empty() {
-        return serde_json::json!({
-            "count": 0,
-            "total": 0,
-            "mean_per_token": serde_json::Value::Null,
-            "max_per_token": serde_json::Value::Null,
-        });
-    }
-
-    let total = samples.iter().map(|sample| sample.alloc_bytes).sum::<u64>();
-    let max = samples.iter().map(|sample| sample.alloc_bytes).max().unwrap_or(0);
-
-    serde_json::json!({
-        "count": samples.len(),
-        "total": total,
-        "mean_per_token": rounded_ms(total as f64 / samples.len() as f64),
-        "max_per_token": max,
-    })
-}
-
-fn mean_alloc_count(samples: &[AllocationAuditSnapshot]) -> Option<f64> {
-    if samples.is_empty() {
-        return None;
-    }
-    Some(samples.iter().map(|sample| sample.alloc_count).sum::<u64>() as f64 / samples.len() as f64)
-}
-
-fn mean_alloc_bytes(samples: &[AllocationAuditSnapshot]) -> Option<f64> {
-    if samples.is_empty() {
-        return None;
-    }
-    Some(samples.iter().map(|sample| sample.alloc_bytes).sum::<u64>() as f64 / samples.len() as f64)
 }
 
 fn percentile_nearest(sorted_samples: &[f64], percentile: usize) -> f64 {
